factor our main for testability
